Nest specific object classes within the mod
