--- osu.Game.Rulesets.Osu/Mods/Objects/StrictTrackingDrawableSliderTail.cs
@@ -1,12 +1,0 @@
-// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-using osu.Game.Rulesets.Osu.Objects.Drawables;
-
-namespace osu.Game.Rulesets.Osu.Mods.Objects
-{
-    public class StrictTrackingDrawableSliderTail : DrawableSliderTail
-    {
-        public override bool DisplayResult => true;
-    }
-}
